UI/Modal/Lightbox: 42416, remove carousel from class
