+++ components/ILIAS/UI/resources/js/Modal/src/Modal.js
@@ -76,75 +76,6 @@ export default class Modal {
     }
     this.#initializedModalboxes[signalData.id] = component.id;
 
-    this.#maybeInitCarousel(component);
-  }
-
-  /**
-   * @param {HTMLDialogElement} component
-   */
-  #maybeInitCarousel(component) {
-    const container = component.querySelector('.carousel-inner');
-    if (!container) {
-      return;
-    }
-
-    const left = component.querySelector('.carousel-control.left');
-    const right = component.querySelector('.carousel-control.right');
-    const indicators = component.querySelectorAll('.carousel-indicators > li');
-
-    left.addEventListener('click', () => this.#nextPage(component, -1));
-    right.addEventListener('click', () => this.#nextPage(component, 1));
-    indicators.forEach(
-      (i) => i.addEventListener(
-        'click',
-        () => this.#gotoPage(component, parseInt(i.getAttribute('data-slide-to'))),
-      ),
-    );
-  }
-
-  #nextPage(component, direction) {
-    const pages = component.querySelectorAll('.carousel-inner > div.item');
-    let index = 0;
-    let current = 0;
-    pages.forEach((p) => {
-      if (p.classList.contains('active')) {
-        current = index;
-      }
-      index += 1;
-    });
-    let next = current + direction;
-    if (next < 0) {
-      next = index - 1;
-    }
-    if (next === index) {
-      next = 0;
-    }
-    this.#gotoPage(component, next);
-  }
-
-  #gotoPage(component, number) {
-    const pages = component.querySelectorAll('.carousel-inner > div.item');
-    const indicators = component.querySelectorAll('.carousel-indicators > li');
-    this.#setActiveInList(pages, number);
-    this.#setActiveInList(indicators, number);
-
-    const title = component.querySelector('h1.modal-title');
-    pages.forEach((item) => {
-      if (item.classList.contains('active')) {
-        title.innerHTML = item.getAttribute('data-title');
-      }
-    });
-  }
-
-  #setActiveInList(list, number) {
-    let index = 0;
-    list.forEach((item) => {
-      if (index !== number) {
-        item.classList.remove('active');
-      } else {
-        item.classList.add('active');
-      }
-      index += 1;
-    });
+    il.UI.lightbox.maybeInitCarousel(component);
   }
 }
